+++ superset-frontend/src/core/sqlLab.ts
@@ -31,18 +31,150 @@ import {
 import { RootState, store } from 'src/views/store';
 import { AnyListenerPredicate } from '@reduxjs/toolkit';
 import type { SqlLabRootState } from 'src/SqlLab/types';
-import {
-  Disposable,
-  Editor,
-  Panel,
-  QueryErrorResultContext,
-  QueryRequestContext,
-  QueryResultContext,
-  Tab,
-} from './core';
+import { Disposable, Editor, Panel, Tab } from './core';
 import { createActionListener } from './utils';
+import { SupersetError } from '@superset-ui/core';
 
 const { CTASMethod } = sqlLabType;
+
+export class CTAS implements sqlLabType.CTAS {
+  method: sqlLabType.CTASMethod;
+
+  tempTable: string;
+
+  constructor(asView: boolean, tempTable: string) {
+    this.method = asView ? CTASMethod.View : CTASMethod.Table;
+    this.tempTable = tempTable;
+  }
+}
+
+export class QueryRequestContext implements sqlLabType.QueryRequestContext {
+  id: string;
+
+  ctas: sqlLabType.CTAS | null;
+
+  editor: Editor;
+
+  queryLimit: number | null;
+
+  runAsync: boolean;
+
+  startDttm: number;
+
+  tab: Tab;
+
+  templateParams: string;
+
+  constructor(
+    id: string,
+    editor: Editor,
+    tab: Tab,
+    runAsync: boolean,
+    startDttm: number,
+    options: {
+      templateParams?: string;
+      ctasMethod?: string;
+      tempTable?: string;
+    } = {},
+  ) {
+    this.id = id;
+    this.editor = editor;
+    this.tab = tab;
+    this.runAsync = runAsync;
+    this.startDttm = startDttm;
+    this.ctas = options.tempTable
+      ? new CTAS(options.ctasMethod === CTASMethod.View, options.tempTable)
+      : null;
+    this.templateParams = options.templateParams ?? '';
+  }
+}
+
+export class QueryResultContext
+  extends QueryRequestContext
+  implements sqlLabType.QueryResultContext
+{
+  endDttm: number;
+
+  executedSql: string;
+
+  limit: number;
+
+  limitingFactor: string;
+
+  queryId: number;
+
+  result: sqlLabType.QueryResult;
+
+  constructor(
+    id: string,
+    queryId: number,
+    executedSql: string,
+    columns: sqlLabType.QueryResult['columns'],
+    data: sqlLabType.QueryResult['data'],
+    editor: Editor,
+    tab: Tab,
+    runAsync: boolean,
+    startDttm: number,
+    endDttm: number,
+    options: {
+      templateParams?: string;
+      ctasMethod?: string;
+      tempTable?: string;
+      limit?: number;
+      limitingFactor?: string;
+    } = {},
+  ) {
+    const { limit, ...opt } = options;
+    super(id, editor, tab, runAsync, startDttm, opt);
+    this.queryId = queryId;
+    this.executedSql = executedSql;
+    this.endDttm = endDttm;
+    this.result = {
+      columns,
+      data,
+    };
+    this.limit = limit ?? data.length;
+    this.limitingFactor = options.limitingFactor ?? '';
+  }
+}
+
+export class QueryErrorResultContext
+  extends QueryRequestContext
+  implements sqlLabType.QueryErrorResultContext
+{
+  endDttm: number;
+
+  errorMessage: string;
+
+  errors: SupersetError[];
+
+  executedSql: string | null;
+
+  constructor(
+    id: string,
+    errorMessage: string,
+    errors: SupersetError[],
+    editor: Editor,
+    tab: Tab,
+    runAsync: boolean,
+    startDttm: number,
+    options: {
+      queryId?: number;
+      executedSql?: string;
+      endDttm?: number;
+      templateParams?: string;
+      ctasMethod?: string;
+      tempTable?: string;
+    } = {},
+  ) {
+    const { queryId, executedSql, endDttm, ...opt } = options;
+    super(id, editor, tab, runAsync, startDttm, opt);
+    this.executedSql = executedSql ?? null;
+    this.errorMessage = errorMessage;
+    this.errors = errors;
+    this.endDttm = endDttm ?? Date.now();
+  }
+}
 
 const activeEditorId = () => {
   const { sqlLab }: { sqlLab: SqlLabRootState['sqlLab'] } = store.getState();
